requirements.txt
fastapi
tensorflow
pillow
numpy
mangum

